Added second html form, refactored shops.
Added a second form to the html that has inputs to change the open and close times of the locations.

Refactored my shop objects into an array and refactored logic to use these objects. Added the code to support the new form and refactored the table builders to clear the current table and build a fresh one.